move components to directory
