+++ javascript/experiments/maze/src/components/App.js
