+++ javascript/experiments/maze/src/components/Canvas.js
