+++ javascript/experiments/maze/src/components/Maze/Edge.js
