+++ javascript/experiments/maze/src/components/Maze/createInnerEdges.js
